% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.index_to_hrf.R
\name{f.index_to_hrf}
\alias{f.index_to_hrf}
\title{Function used to determine hrf from index given experiment parameters}
\usage{
f.index_to_hrf(index, width, dist.bas, Zf, data)
}
\arguments{
\item{index}{index}

\item{width}{width of the plate}

\item{dist.bas}{dist.bas}

\item{Zf}{Zf}

\item{data}{image of chromatograms}
}
\description{
Function used to determine hrf from index given experiment parameters
}
\author{
Dimitri Fichou
}
