% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.array.R
\name{f.plot.array}
\alias{f.plot.array}
\title{Function to plot the video densitograms return by f.eat.image}
\usage{
f.plot.array(data, id, hauteur, Zf, dist.bas, reconstruct = T, inverse = F,
  ylim.raster = 1.3, main = NULL, ...)
}
\arguments{
\item{data}{input array or matrix}

\item{id}{index of the track}

\item{hauteur}{plate width}

\item{Zf}{migration front}

\item{dist.bas}{distance to lower edge}

\item{reconstruct}{should raster be plotted}

\item{inverse}{legacy from rtlc where the xlim were inversed on mac}

\item{ylim.raster}{size of the raster if applicable}

\item{...}{extra parameters to be passed to the plot function}
}
\description{
Function to plot the video densitograms return by f.eat.image
}
\author{
Dimitri Fichou
}
