% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.function.R
\name{f.preprocess}
\alias{f.preprocess}
\title{function to do the full preprocess}
\usage{
f.preprocess(data, preprocess.order, preprocess.option, training.data = data)
}
\arguments{
\item{data}{an array, or a matrix}

\item{preprocess.order}{preprocess.order}

\item{preprocess.option}{preprocess.option}

\item{training.data}{training.data}

\item{...}{Arguments to be passed to methods}
}
\description{
function to do the full preprocess
}
\author{
Dimitri Fichou
}
