% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track.R
\name{plot_track}
\alias{plot_track}
\title{Function to plot the 3 channels of an array as chromatograms,}
\usage{
plot_track(data, normalization = T, ...)
}
\arguments{
\item{data}{a 3D array of dim[3] = 3}

\item{normalization}{either or not normalisation should happen before plotting the chromatograms}

\item{...}{extra parameters to be passed to the plot function}

\item{x}{the x coordinate to look at}

\item{edge}{to apply a mean and avoid noise}
}
\description{
Function to plot the 3 channels of an array as chromatograms,
}
\author{
Dimitri Fichou
}
