% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.array.R
\name{f.plot.array}
\alias{f.plot.array}
\title{function to plot an array}
\usage{
f.plot.array(
  data,
  id,
  hauteur,
  Zf,
  dist.bas,
  reconstruct = T,
  ylim.raster = 1.3,
  main = NULL,
  xlim = NULL,
  channel = NULL,
  ...
)
}
\arguments{
\item{data}{an array, or a matrix}

\item{id}{which row of the array to plot}

\item{hauteur}{width of the plate to put correct hrf}

\item{Zf}{Zf to put correct hrf}

\item{dist.bas}{application distance to put correct hrf}

\item{reconstruct}{option to reconstruct or not the image}

\item{ylim.raster}{optin to control the width of the image to reconstruct}

\item{main}{optionnal title}

\item{xlim}{optionnal xlim to zoom in}

\item{channel}{plot a single channel}

\item{...}{Arguments to be passed to methods}
}
\description{
function to plot an array
}
\author{
Dimitri Fichou
}
